+++ crates/sui-core/src/authority_server.rs
@@ -25,8 +25,7 @@ use sui_types::messages_grpc::{
 };
 use sui_types::multiaddr::Multiaddr;
 use sui_types::sui_system_state::SuiSystemState;
-use sui_types::traffic_control::ServiceResponse;
-use sui_types::traffic_control::{PolicyConfig, RemoteFirewallConfig};
+use sui_types::traffic_control::{PolicyConfig, RemoteFirewallConfig, Weight};
 use sui_types::{error::*, transaction::*};
 use sui_types::{
     fp_ensure,
@@ -659,7 +658,9 @@ impl ValidatorService {
         proxy_ip: Option<SocketAddr>,
     ) -> Result<(), tonic::Status> {
         if let Some(traffic_controller) = &self.traffic_controller {
-            if !traffic_controller.check(connection_ip, proxy_ip).await {
+            let connection = connection_ip.map(|ip| ip.ip());
+            let proxy = proxy_ip.map(|ip| ip.ip());
+            if !traffic_controller.check(connection, proxy).await {
                 // Entity in blocklist
                 if traffic_controller.dry_run_mode() {
                     debug!(
@@ -685,17 +686,17 @@ impl ValidatorService {
         proxy_ip: Option<SocketAddr>,
         response: &Result<tonic::Response<T>, tonic::Status>,
     ) {
-        let result: SuiResult = if let Err(status) = response {
-            Err(SuiError::from(status.clone()))
+        let error: Option<SuiError> = if let Err(status) = response {
+            Some(SuiError::from(status.clone()))
         } else {
-            Ok(())
+            None
         };
 
         if let Some(traffic_controller) = self.traffic_controller.clone() {
             traffic_controller.tally(TrafficTally {
                 connection_ip: connection_ip.map(|ip| ip.ip()),
                 proxy_ip: proxy_ip.map(|ip| ip.ip()),
-                result: ServiceResponse::Validator(result),
+                weight: error.map(normalize).unwrap_or(Weight::zero()),
                 timestamp: SystemTime::now(),
             })
         }
@@ -712,6 +713,20 @@ fn make_tonic_request_for_testing<T>(message: T) -> tonic::Request<T> {
     };
     request.extensions_mut().insert(tcp_connect_info);
     request
+}
+
+// TODO: refine error matching here
+fn normalize(err: SuiError) -> Weight {
+    match err {
+        SuiError::UserInputError { .. }
+        | SuiError::InvalidSignature { .. }
+        | SuiError::SignerSignatureAbsent { .. }
+        | SuiError::SignerSignatureNumberMismatch { .. }
+        | SuiError::IncorrectSigner { .. }
+        | SuiError::UnknownSigner { .. }
+        | SuiError::WrongEpoch { .. } => Weight::one(),
+        _ => Weight::zero(),
+    }
 }
 
 /// Implements generic pre- and post-processing. Since this is on the critical
@@ -731,7 +746,9 @@ macro_rules! handle_with_decoration {
         // is hitting this case, we should reject such requests that
         // hit this case.
         if connection_ip.is_none() {
-            if cfg!(all(test, not(msim))) {
+            if cfg!(msim) {
+                // Ignore the error from simtests.
+            } else if cfg!(test) {
                 panic!("Failed to get remote address from request");
             } else {
                 $self.metrics.connection_ip_not_found.inc();