+++ sdk/enoki/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mysten/enoki
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [e279098cff]
+  - @mysten/zklogin@0.6.0
+
 ## 0.2.3
 
 ### Patch Changes